--- a/db/view/view_update_generator.cc
+++ b/db/view/view_update_generator.cc
@@ -462,6 +462,7 @@ future<> view_update_generator::generate_and_propagate_view_updates(const replic
         }
     }
     co_await builder.close();
+    _proxy.local().update_view_update_backlog();
     if (err) {
         std::rethrow_exception(err);
     }