This series is another approach of https://github.com/scylladb/scylladb/pull/18646. The main difference
between the two is that in the former series the updated backlog was propagated from the view updating
code to the write request handling using return values on a few different layers. This series performs
the atomic read of the view update backlog from the correct shard on the shard handling the replica
write request.
This series should be simpler thanks to the recent separation of backlog update and backlog retrieval,
however this can also benefit the original series after it's updated. It also modifies fewer layers, which
makes the series more concise.
The downside of this approach is the cross-shard read of an atomic holding the view update backlog of
another shard, which was performed inside the `invoke_on` in the original series, and the fact that
this read is now done regardless of whether the write actually generates view updates - we can only
discover this when we start processing the mutation. Additionally, the shard calculation also needed to
be duplicated - currently it's done in `storage_proxy::mutate_locally` where we can't add the view backlog
retrieval without adding an extra allocation to the common write path.
Due to this extra work, the instruction count per operation increased notably in the `perf-simple-query`
results, to an extend similar of adding the extra allocation to `storage_proxy::mutate_locally` - the results
are compared at the end of the cover message.

For now, this patch doesn't include getting the backlog atomic on failed requests - this would require
quite a few extra modifications only giving a small improvement for the uncommon patch - for now
it's left as a FIXME

The detailed description of the problem is mostly the same as in the original series, included below:

When a replica applies a write on a table which has a materialized view
it generates view updates. These updates take memory which is tracked
by `database::_view_update_concurrency_sem`, separate on each shard.
The fraction of units taken from the semaphore to the semaphore limit
is the shard's view update backlog. Based on these backlogs, we want
to estimate how busy a node is with its view updates work. We do that
by taking the max backlog across all shards.
To avoid excessive cross-shard operations, the node's (max) backlog isn't
calculated each time we need it, but up to 1 time per 10ms (the `_interval`) with an optimization where the backlog of the calculating shard is immediately up-to-date (we don't need cross-shard operations for it):
```
update_backlog node_update_backlog::fetch() {
    auto now = clock::now();
    if (now >= _last_update.load(std::memory_order_relaxed) + _interval) {
        _last_update.store(now, std::memory_order_relaxed);
        auto new_max = boost::accumulate(
                _backlogs,
                update_backlog::no_backlog(),
                [] (const update_backlog& lhs, const per_shard_backlog& rhs) {
                    return std::max(lhs, rhs.load());
                });
        _max.store(new_max, std::memory_order_relaxed);
        return new_max;
    }
    return std::max(fetch_shard(this_shard_id()), _max.load(std::memory_order_relaxed));
}
```
For the same reason, even when we do calculate the new node's backlog,
we don't read from the `_view_update_concurrency_sem`. Instead, for
each shard we also store a update_backlog atomic which we use for
calculation:
```
    struct per_shard_backlog {
        // Multiply by 2 to defeat the prefetcher
        alignas(seastar::cache_line_size * 2) std::atomic<update_backlog> backlog = update_backlog::no_backlog();
        need_publishing need_publishing = need_publishing::no;

        update_backlog load() const {
            return backlog.load(std::memory_order_relaxed);
        }
    };
 std::vector<per_shard_backlog> _backlogs; 
```
Due to this distinction, the update_backlog atomic need to be updated
separately, when the `_view_update_concurrency_sem` changes.
This is done by calling `storage_proxy::update_view_update_backlog`, which reads the `_view_update_concurrency_sem` of the shard (in `database::get_view_update_backlog`)
and then calls node`_update_backlog::add` where the read backlog
is stored in the atomic:
```
void storage_proxy::update_view_update_backlog() {
    _max_view_update_backlog.add(get_db().local().get_view_update_backlog());
}
void node_update_backlog::add(update_backlog backlog) {
    _backlogs[this_shard_id()].backlog.store(backlog, std::memory_order_relaxed);
    _backlogs[this_shard_id()].need_publishing = need_publishing::yes;
}
```
For this implementation of calculating the node's view update backlog to work,
we need the atomics to be updated correctly when the semaphores of corresponding
shards change.

The main event where the view update backlog changes is an incoming write
request. That's why when handling the request and preparing a response
we update the backlog calling `storage_proxy::get_view_update_backlog` (also
because we want to read the backlog and send it in the response):
backlog update after local view updates (`storage_proxy::send_to_live_endpoints` in `mutate_begin`)
```
 auto lmutate = [handler_ptr, response_id, this, my_address, timeout] () mutable { 
     return handler_ptr->apply_locally(timeout, handler_ptr->get_trace_state()) 
             .then([response_id, this, my_address, h = std::move(handler_ptr), p = shared_from_this()] { 
         // make mutation alive until it is processed locally, otherwise it 
         // may disappear if write timeouts before this future is ready 
         got_response(response_id, my_address, get_view_update_backlog()); 
     }); 
 }; 
backlog update after remote view updates (storage_proxy::remote::handle_write)

 auto f = co_await coroutine::as_future(send_mutation_done(netw::messaging_service::msg_addr{reply_to, shard}, trace_state_ptr, 
         shard, response_id, p->get_view_update_backlog())); 
```
Now assume that on a certain node we have a write request received on shard A,
which updates a row on shard B (A!=B). As a result, shard B will generate view
updates and consume units from its `_view_update_concurrency_sem`, but will
not update its atomic in `_backlogs` yet. Because both shards in the example
are on the same node, shard A will perform a local write calling `lmutate` shown
above. In the `lmutate` call, the `apply_locally` will initiate the actual write on
shard B and the `storage_proxy::update_view_update_backlog` will be called back
on shard A. In no place will the backlog atomic on shard B get updated even
though it increased in size due to the view updates generated there.
Currently, what we calculate there doesn't really matter - it's only used for the
MV flow control delays, so currently, in this scenario, we may only overload
a replica causing failed replica writes which will be later retried as hints. However,
when we add MV admission control, the calculated backlog will be the difference
between an accepted and a rejected request.

The fix for this correctness issue we need to update the view update backlog
on the shard whose semaphore units have been consumed. We could still read the
backlog on the shard that received the initial request, however, if we read the backlog
on a different shard than the one which generated the view updates, it will most likely
be outdated (by 10ms), so we'll likely won't see the backlog change caused by the write
request. If we want to see that, we need to get the return updated value from the shard
that generated the view updates and return it on the shard handling the request.

This patch ensures that when semaphore units are consumed on some shard, it's
atomic backlog is updated and that when a backlogs increases as a result of some
request, the increased backlog is returned in the response of this request. This
is achieved by reading the view update backlog atomic holding the backlog from
the shard where the mutation was applied when finishing the write request handling.
The exact place where we get the backlog was chosen due to performance considerations:
we don't want to create a new task on the common write path just for this, so it's inside the
`lmutate` and `handle_write` coroutines.

Perf simple query in write mode gives the following results:
Before:
median 337720.61 tps ( 59.3 allocs/op,  16.0 logallocs/op,  15.0 tasks/op,   52842 insns/op,        0 errors)
median absolute deviation: 1848.82
maximum: 342161.04
minimum: 299399.57
After:
median 333203.07 tps ( 59.3 allocs/op,  16.0 logallocs/op,  15.0 tasks/op,   53585 insns/op,        0 errors)
median absolute deviation: 5629.36
maximum: 339708.32
minimum: 309021.10

Fixes: https://github.com/scylladb/scylladb/issues/18542

The test added in this patch assures that the backlog is updated on the correct shard. The value
returned in responses isn't currently used for anything other than delaying responses, so it will
be tested when it start having an impact in https://github.com/scylladb/scylladb/pull/18334.

Without admission control (https://github.com/scylladb/scylladb/pull/18334), this patch doesn't affect much, so I'm marking it as backport/none