@nyh 2024-06-09
> This series is another approach of #18646.

Both PRs are open now, so what is your intention? Do you prefer this one over the other once (in that case, please closed the other one or at least change it into a "draft" so nobody considers merging it)? Or do you want the reviewers to tell you which one is better?