@wmitros 2024-06-10
> > This series is another approach of #18646.
> 
> Both PRs are open now, so what is your intention? Do you prefer this one over the other once (in that case, please closed the other one or at least change it into a "draft" so nobody considers merging it)? Or do you want the reviewers to tell you which one is better?
 
I mostly need this issue fixed for https://github.com/scylladb/scylladb/pull/18334 (the update part is needed for it to work at all, and the returning part is mostly for its testing), so I'm fine with either of these two PRs as they both fix the problem. I added this second PR because the previous one didn't make any progress with the review and the new approach was suggested in https://github.com/scylladb/scylladb/pull/18646#discussion_r1604768817.

Personally, I still prefer the original solution as it doesn't affect the performance of the common write path and it's impact on the update-generating path is also smaller, even though it affects many more seemingly unrelated components. I'm hoping after that seeing both possible implementations we'll be able to choose at least one that we can go forward with.

There's also a third, least complex, approach that I can prepare, but it adds an extra allocation to the write path, which I assume we want to avoid