@wmitros 2024-06-12
> There's also a third, least complex, approach that I can prepare, but it adds an extra allocation to the write path, which I assume we want to avoid

Actually, I take that back. The approach that adds the extra allocation is almost as widespread as https://github.com/scylladb/scylladb/pull/18646 (we still need to return the backlog in return values of `apply_locally` and `modify_locally`) so it doesn't actually have any advantages.