@wmitros 2024-06-12
@avikivity @gleb-cloudius you'd be welcome to view the mv changes, but I'd like to primarily ask you what do you think about the performance impact of this change and compare it to the added complexity in https://github.com/scylladb/scylladb/pull/18646 (the second approach for this issue).

The problem is described extensively in the cover letter, so the following description may be not clear but trying to be shorter:
For each shard of a node we keep its view update backlog (the memory used by currently processed view updates), which are aggregated and their max is the node's backlog, and this max is sent to other nodes to be used for the mv flow control (delaying responses depending on the mv loads).

The first issue is with the aggregation, but it's a relatively small change - fixed in commit https://github.com/scylladb/scylladb/pull/19181/commits/202392cf752094dde9f9e1f5bf79fce9b4db5f69. With this change alone the main issue of backlog changes just not appearing will be fixed.

The second issue is with the sending of the backlog. Currently, we send it in two ways - every 1s using gossip and adding the backlog to replica write responses. Because during the aggregation we read from atomics for all shards, we only perform it once every 10ms (or more rarely). As a result, with just the first fix, the response may not contain the backlog increase that happened because of the exact write that we're responding to. The current implementation intended to improve this situation by using the max of aggregated backlog and the backlog from the shard where view updates were generated, but instead it got the max of aggregated backlog and the backlog from the shard which received the replica write, not necessarily applied it.
The two approaches (this PR and https://github.com/scylladb/scylladb/pull/18646) differ mainly in how they fix the second issue.
In this one, on the shard which received the replica write, we perform a cross shard read to the atomic from the shard were view backlog increased. We perform the read not knowing whether the write is generating any view updates, so it adds extra instructions to the common write path.
In https://github.com/scylladb/scylladb/pull/18646, we also read from the atomic, but we do it directly around the code where view updates are generated, and the backlog is then propagated through the call stack (`sp::apply_locally`-`sp::mutate_locally`-`db::apply`-`cf::push_view_replica_updates`-`vug::generate_and_propagate_view_updates`) in return values. This approach has the benefit of not affecting the common write path but it modifies a number of seemingly unrelated components.